+++ include/universal/number/lns/lns_traits.hpp
@@ -15,7 +15,7 @@ struct is_lns_trait
 {
 };
 
-template<size_t nbits, size_t rbits, typename BlockType, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename BlockType, auto... xtra>
 struct is_lns_trait< lns<nbits, rbits, BlockType, xtra...> >
 	: true_type
 {
