+++ include/universal/number/lns/math/complex.hpp
@@ -9,21 +9,21 @@
 namespace sw { namespace universal {
 
 // Real component of a complex lns
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...>
 real(std::complex< lns<nbits, rbits, bt, xtra...> > x) {
 	return lns<nbits, rbits, bt, xtra...>(std::real(x));
 }
 
 // Imaginary component of a complex lns
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...>
 imag(std::complex< lns<nbits, rbits, bt, xtra...> > x) {
 	return lns<nbits, rbits, bt, xtra...>(std::imag(x));
 }
 
 // Conjucate of a complex lns
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 std::complex< lns<nbits, rbits, bt, xtra...> >
 conj(std::complex< lns<nbits, rbits, bt, xtra...> > x) {
 	return lns<nbits, rbits, bt, xtra...>(std::conj(x));
